runtime.WebhookProvider — generic webhook routes from credential plugins
main used to hardcode /api/slack/interactive in newWebMux. Replace
with a runtime interface: any credential body that implements
WebhookProvider gets its routes mounted under /api/cred/<credName>/
at boot. Slack's interactive callback uses this; future Discord /
Telegram / generic-webhook plugins plug in by implementing the same
method.

  // config/runtime/webhook.go
  type WebhookProvider interface {
      WebhookRoutes() []WebhookRoute
  }
  type WebhookRoute struct {
      Path    string
      Handler WebhookHandler
  }
  type WebhookCtx struct {
      CredentialName string
      Secrets        SecretStore
      HITL           HITLPool
      Policy         *config.CompiledPolicy
      Profiles       []string
  }

HITLPool gains Decide so a webhook can forward a side-channel verdict
into the same pool the dashboard writes to (HITLRegistry already
implements it).

slack_interactive.go (top-level) folds into config/plugins/credentials/slack.go
— signature verify + payload parse + form decode + response_url POST
all live next to the SlackTokens struct now.

Public path moves from /api/slack/interactive to
/api/cred/<credName>/interactive — operator-visible breaking change.
Update Slack app's "Interactivity & Shortcuts" URL accordingly.
